Tests for commment dao

+++ blablacar-persistence/src/main/java/cz/muni/fi/pa165/blablacar/persistence/dao/UserDaoImpl.java
@@ -28,26 +28,44 @@ public class UserDaoImpl implements UserDao {
 
     @Override
     public void addUser(User u) throws IllegalArgumentException {
+        if(u == null){
+            throw new IllegalArgumentException("User is null");
+        }
         em.persist(u);
     }
 
     @Override
-    public void updateUser(User u) {
+    public void updateUser(User u) throws IllegalArgumentException{
+        if(u == null){
+            throw new IllegalArgumentException("User is null");
+        }
         em.merge(u);
     }
 
     @Override
-    public void removeUser(User u) {
+    public void removeUser(User u) throws IllegalArgumentException{
+        if(u == null){
+            throw new IllegalArgumentException("User is null");
+        }
         em.remove(u);
     }
 
     @Override
-    public User findUserById(Long id) {
+    public User findUserById(Long id) throws IllegalArgumentException{
+        if(id == null){
+            throw new IllegalArgumentException("Id is null");
+        }
         return em.find(User.class, id);
     }
 
     @Override
-    public User findUserByFullName(String fName, String lName) {
+    public User findUserByFullName(String fName, String lName) throws IllegalArgumentException{
+        if(fName == null){
+            throw new IllegalArgumentException("First name is null");
+        }
+        if(lName == null){
+            throw new IllegalArgumentException("Last name is null");
+        }
         try {
             return em.createQuery("select u from User u where firstName = :fName and lastName = :lName", User.class)
                     .setParameter("fName", fName)
@@ -59,7 +77,10 @@ public class UserDaoImpl implements UserDao {
     }
 
     @Override
-    public User findUserByLogin(String login) {
+    public User findUserByLogin(String login) throws IllegalArgumentException{
+        if(login == null){
+            throw new IllegalArgumentException("Login is null");
+        }
         try {
             return em.createQuery("select u from User u where login = :login", User.class)
                     .setParameter("login", login)
